# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Euaconecta Platform API
  description: API RESTful para gerenciamento de pacotes, consolidações e operações logísticas
  version: 1.0.0
  contact:
    name: Suporte Euaconecta
    email: suporte@euaconecta.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Servidor de desenvolvimento
  - url: https://api.euaconecta.com
    description: Servidor de produção
security:
  - sessionAuth: []
paths:
  /api/auth/login:
    post:
      tags:
        - Autenticação
      summary: Autentica um usuário
      description: Autentica um usuário no sistema e retorna token de sessão
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/packages:
    get:
      tags:
        - Pacotes
      summary: Lista pacotes do usuário
      description: Retorna lista de pacotes do usuário autenticado
      parameters:
        - name: status
          in: query
          description: Filtrar por status
          schema:
            type: string
            enum: [RECEIVED, CONFIRMED, IN_CONSOLIDATION, CONSOLIDATED, SHIPPED]
        - name: page
          in: query
          description: Número da página
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Itens por página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Lista de pacotes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Package'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Pacotes
      summary: Cria um novo pacote
      description: Cria um novo pacote para o usuário autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePackageRequest'
      responses:
        '201':
          description: Pacote criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Package'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/freight/calculate:
    get:
      tags:
        - Frete
      summary: Calcula frete
      description: Calcula o valor do frete para um envio
      parameters:
        - name: weight
          in: query
          required: true
          description: Peso em gramas
          schema:
            type: integer
            minimum: 1
            example: 1000
        - name: destination
          in: query
          required: true
          description: Destino
          schema:
            type: string
            example: SP
        - name: serviceType
          in: query
          description: Tipo de serviço
          schema:
            type: string
            enum: [STANDARD, EXPRESS, OVERNIGHT]
            default: STANDARD
      responses:
        '200':
          description: Cálculo do frete
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      cost:
                        type: number
                        format: float
                        example: 15.99
                      currency:
                        type: string
                        example: USD
                      estimatedDays:
                        type: integer
                        example: 3
                      carrier:
                        type: string
                        example: UPS
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Token de sessão obtido através do login
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: cmfjo89ni0001omkkxndmuu77
        name:
          type: string
          example: João Silva
        email:
          type: string
          format: email
          example: joao@example.com
        role:
          type: string
          enum: [CLIENT, ADMIN]
          example: CLIENT
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
    Package:
      type: object
      properties:
        id:
          type: string
          example: cmfjo89ni0001omkkxndmuu77
        description:
          type: string
          example: iPhone 15 Pro
        weightGrams:
          type: integer
          example: 187
        purchasePrice:
          type: number
          format: float
          example: 999.99
        store:
          type: string
          example: Apple Store
        orderNumber:
          type: string
          example: APL123456789
        carrier:
          type: string
          example: UPS
        declaredValue:
          type: number
          format: float
          example: 999.99
        packageType:
          type: string
          enum: [BOX, ENVELOPE, PACKAGE]
          example: BOX
        status:
          type: string
          enum: [RECEIVED, CONFIRMED, IN_CONSOLIDATION, CONSOLIDATED, SHIPPED]
          example: RECEIVED
        lengthCm:
          type: integer
          example: 15
        widthCm:
          type: integer
          example: 8
        heightCm:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
    CreatePackageRequest:
      type: object
      required:
        - description
        - weightGrams
        - purchasePrice
        - store
        - orderNumber
        - carrier
        - declaredValue
        - packageType
        - lengthCm
        - widthCm
        - heightCm
      properties:
        description:
          type: string
          example: iPhone 15 Pro
        weightGrams:
          type: integer
          minimum: 1
          example: 187
        purchasePrice:
          type: number
          format: float
          minimum: 0
          example: 999.99
        store:
          type: string
          example: Apple Store
        orderNumber:
          type: string
          example: APL123456789
        carrier:
          type: string
          example: UPS
        declaredValue:
          type: number
          format: float
          minimum: 0
          example: 999.99
        packageType:
          type: string
          enum: [BOX, ENVELOPE, PACKAGE]
          example: BOX
        lengthCm:
          type: integer
          minimum: 1
          example: 15
        widthCm:
          type: integer
          minimum: 1
          example: 8
        heightCm:
          type: integer
          minimum: 1
          example: 1
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Mensagem de erro específica
        code:
          type: string
          example: ERROR_CODE
        details:
          type: object
          example: {}
...